Move notification title evaluation to notification.service so it can be reused
